pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "bifrost-trader-engine"
version = "0.1.0"
description = "Gamma scalping trading daemon for NVDA 21-35 DTE near-ATM straddle (IB)"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "ib_insync>=0.9.86",
    "psycopg2-binary>=2.9.0",
    "py_vollib>=1.0.1",
    "PyYAML>=6.0",
    "fastapi>=0.100.0",
    "uvicorn[standard]>=0.22.0",
    "pytest>=7.0",
    "pytest-asyncio>=0.21",
]

[project.optional-dependencies]
dev = ["pytest>=7.0", "pytest-asyncio>=0.21"]
docs = [
    "mkdocs>=1.5.0",
    "mkdocs-material>=9.0.0",
    "mkdocs-mermaid2-plugin>=1.0.0",
    "pymdown-extensions>=10.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["src*", "config*"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
addopts = "-v"
markers = [
    "ib: mark test as requiring live IB connection (deselect with -m 'not ib')",
]
